+++ CHANGELOG.md
@@ -12,18 +12,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TableViewModel for admin.latte
 - nav.latte is used in the BlueprintWeb.latte and can be used in app (using inherite.latte)
 - blast.sh self-update (from the Seablast library)
+- assets/scripts expected by plugins (Seablast/Auth) send-auth-token.js expecting both Environment.js and seablast.js
 
 ### `Changed` for changes in existing functionality
 
 - better error reporting in Apis\GenericRestApiJsonModel
 - blast.sh main requires confirmation
 - PHPUnit tests folder renamed according to a common convention
+- GenericRestApiJsinModel.php is more verbose about an input JSON error
 
 ### `Deprecated` for soon-to-be removed features
 
 ### `Removed` for now removed features
 
-- phpstan.sh removed as blast.sh can be used instead
+- phpstan.sh removed because blast.sh can be used instead
 
 ### `Fixed` for any bugfixes